Moving GSC models to models_extra
