+++ docs/reference_guides/model_libraries/gas_solid_contactors/flowsheets/ss_BFB_methane_combustion.rst
@@ -1,7 +1,7 @@
 Methane Combustion in a Bubbling Fluidized Bed (steady-state)
 =============================================================
 
-.. currentmodule:: idaes.gas_solid_contactors.flowsheets.ss_BFB_methane_combustion
+.. currentmodule:: idaes.models_extra.gas_solid_contactors.flowsheets.ss_BFB_methane_combustion
 
 Steady-state flowsheet example of the bubbling fluidized bed model for methane combustion with an iron-oxide based oxygen carrier.
