+++ docs/reference_guides/model_libraries/gas_solid_contactors/properties/methane_iron_OC_reduction/heterogeneous_reaction_properties.rst
@@ -2,12 +2,12 @@ Heterogeneous reaction properties
 =================================
 
 .. index::
-   pair: idaes.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionParameterBlock
-   pair: idaes.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionParameterData
-   pair: idaes.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionStateBlock
-   pair: idaes.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionStateBlockData
+   pair: idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionParameterBlock
+   pair: idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionParameterData
+   pair: idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionStateBlock
+   pair: idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions;HeteroReactionStateBlockData
 
-.. currentmodule:: idaes.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions
+.. currentmodule:: idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction.hetero_reactions
 
 Property package for the reaction of CH4 with an iron-based OC. More details of this reaction scheme can be found in reference [1].
 The gas components modeled are CH4, CO2, H2O.
