+++ idaes/models_extra/gas_solid_contactors/properties/methane_iron_OC_reduction/hetero_reactions.py
