+++ idaes/models_extra/gas_solid_contactors/properties/methane_iron_OC_reduction/solid_phase_thermo.py
