+++ idaes/models_extra/gas_solid_contactors/properties/oxygen_iron_OC_oxidation/tests/test_CLC_rxn_prop.py
@@ -43,11 +43,11 @@ from idaes.core.util.model_statistics import (
 
 from idaes.core.solvers import get_solver
 
-from idaes.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
+from idaes.models_extra.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
     gas_phase_thermo import GasPhaseParameterBlock
-from idaes.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
+from idaes.models_extra.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
     solid_phase_thermo import SolidPhaseParameterBlock
-from idaes.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
+from idaes.models_extra.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
     hetero_reactions import HeteroReactionParameterBlock
