+++ idaes/models_extra/gas_solid_contactors/properties/oxygen_iron_OC_oxidation/tests/test_CLC_solid_prop.py
@@ -43,7 +43,7 @@ from idaes.core.util.model_statistics import (
 
 from idaes.core.solvers import get_solver
 
-from idaes.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
+from idaes.models_extra.gas_solid_contactors.properties.oxygen_iron_OC_oxidation. \
     solid_phase_thermo import SolidPhaseParameterBlock
 
 # Get default solver for testing
