+++ idaes/models_extra/gas_solid_contactors/unit_models/__init__.py
@@ -10,6 +10,6 @@
 # Please see the files COPYRIGHT.md and LICENSE.md for full copyright and
 # license information.
 #################################################################################
-from idaes.gas_solid_contactors.unit_models.bubbling_fluidized_bed import BubblingFluidizedBed
-from idaes.gas_solid_contactors.unit_models.moving_bed import MBR
-from idaes.gas_solid_contactors.unit_models.fixed_bed_0D import FixedBed0D
+from idaes.models_extra.gas_solid_contactors.unit_models.bubbling_fluidized_bed import BubblingFluidizedBed
+from idaes.models_extra.gas_solid_contactors.unit_models.moving_bed import MBR
+from idaes.models_extra.gas_solid_contactors.unit_models.fixed_bed_0D import FixedBed0D
