+++ idaes/models_extra/gas_solid_contactors/unit_models/bubbling_fluidized_bed.py
