+++ idaes/models_extra/gas_solid_contactors/unit_models/fixed_bed_0D.py
