+++ idaes/models_extra/gas_solid_contactors/unit_models/moving_bed.py
