+++ idaes/models_extra/gas_solid_contactors/unit_models/tests/test_BFB.py
@@ -42,13 +42,13 @@ from idaes.core.solvers import get_solver
 from idaes.core.util import scaling as iscale
 from idaes.core.util.exceptions import InitializationError
 
-from idaes.gas_solid_contactors.unit_models. \
+from idaes.models_extra.gas_solid_contactors.unit_models. \
     bubbling_fluidized_bed import BubblingFluidizedBed
-from idaes.gas_solid_contactors.properties.methane_iron_OC_reduction. \
+from idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction. \
     gas_phase_thermo import GasPhaseParameterBlock
-from idaes.gas_solid_contactors.properties.methane_iron_OC_reduction. \
+from idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction. \
     solid_phase_thermo import SolidPhaseParameterBlock
-from idaes.gas_solid_contactors.properties.methane_iron_OC_reduction. \
+from idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction. \
     hetero_reactions import HeteroReactionParameterBlock
 
 # -----------------------------------------------------------------------------
