+++ idaes/models_extra/gas_solid_contactors/unit_models/tests/test_FB0D.py
@@ -37,14 +37,14 @@ from idaes.core.util.initialization import initialize_by_time_element
 from idaes.core.solvers import get_solver
 
 # Import FixedBed0D unit model
-from idaes.gas_solid_contactors.unit_models.fixed_bed_0D import FixedBed0D
+from idaes.models_extra.gas_solid_contactors.unit_models.fixed_bed_0D import FixedBed0D
 
 # Import property packages
-from idaes.gas_solid_contactors.properties.methane_iron_OC_reduction. \
+from idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction. \
     gas_phase_thermo import GasPhaseParameterBlock
-from idaes.gas_solid_contactors.properties.methane_iron_OC_reduction. \
+from idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction. \
     solid_phase_thermo import SolidPhaseParameterBlock
-from idaes.gas_solid_contactors.properties.methane_iron_OC_reduction. \
+from idaes.models_extra.gas_solid_contactors.properties.methane_iron_OC_reduction. \
     hetero_reactions import HeteroReactionParameterBlock
 
 # -----------------------------------------------------------------------------
